old-tests/submission/Microsoft/forms/formmethod.html

<!DOCTYPE html>
<html><head>
        <title>formmethod on button element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="content-type">
        <meta content="formmethod" name="description">
    <link href="http://www.w3.org/TR/html5/attributes-common-to-form-controls.html#dom-fs-formmethod" rel="help">
</head>
    <body>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <div id="log"></div>
        <button formmethod="post" style="display: none" type="submit">Submit</button>
        <input formmethod="post" style="display: none" type="submit" value="submit">
        <input formmethod="get" style="display: none" type="submit" value="submit">
        
        
        <script type="text/javascript">
        
            test(function() {assert_true("formMethod" in document.getElementsByTagName("button")[0])}, "formMethod attribute 'in' support fon button element");
            test(function() {assert_equals(document.getElementsByTagName("button")[0].formMethod.toLowerCase(), "post")}, "formMethod returns post for button element");

            test(function() {assert_true("formMethod" in document.getElementsByTagName("input")[0])}, "formMethod attribute 'in' support on input element");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].formMethod.toLowerCase(), "post")}, "formMethod returns post for input element");

            var testElem = document.getElementsByTagName("input")[1];
            testElem.formMethod = "post";
            
            test(function() {assert_true("formMethod" in document.getElementsByTagName("input")[1])}, "formMethod attribute 'in' support on input element exsists after modification");
            test(function() {assert_equals(document.getElementsByTagName("input")[1].formMethod.toLowerCase(), "post")}, "formMethod returns correct value after being changed by the DOM");
            
        </script>
    
</body></html>
